requirements.txt
fastapi
uvicorn
pillow
rembg
ultralytics
requests
numpy
pandas
colorthief
cloudinary
onnx
onnxruntime
requests
python-dotenv
webcolors
pydantic
json


